chore: refactor patientdashboard a make linter better
