+++ .flake8
@@ -1,2 +1,5 @@
 [flake8]
 max-line-length = 140
+exclude =
+    venv
+    node_modules
